requirements.txt
openai-whisper
sounddevice
pynput
pystray
pillow
